01/11/2027 19:05

+++ src/main/java/com/ccc/parcauto/detail_besoin/model/DetailBesoins.java
@@ -21,7 +21,7 @@ import java.math.BigInteger;
 @Entity
 public class DetailBesoins extends AbstractEntity {
 private String description;
-private Integer qte;
+private BigInteger qte;
 private BigInteger prix_u;
 private BigInteger totalBesoins;
